pdf_enhancer_requirements.txt
torch>=2.0.0
torchvision>=0.15.0
PyMuPDF>=1.23.0
Real-ESRGAN>=0.3.0
basicsr>=1.4.2
gfpgan>=1.3.8
opencv-python>=4.8.0
psutil>=5.9.0
GPUtil>=1.4.0
tqdm>=4.65.0
Pillow>=10.0.0
numpy>=1.24.0